Update TOCInline.tsx
Enhances functionality by providing users with the option to customize the styling of list (li) elements. I'm currently utilizing this library and felt the need to personalize the appearance of my li elements. Therefore, I'm suggesting this modification to the library to ensure that everyone can take advantage of this improvement.